requirements.txt
python-telegram-bot
asyncio
pathlib
typing
dataclasses
collections
datetime
logging
json
random
math
os